Add type casting operators
- <expr> as? T (returns Optional<T>)
- <expr> as! T (returns T or panics)
- <expr> as saturated T (returns T, value saturated to min/max)
- <expr> as truncated T (returns T, value truncated to fit)

Note that `as?` will return None if the original value does not fit in
the new type. `as!` will panic for the same reason.

The `as saturated` and `as truncated` casts are useful when casting
numeric values, and give the user full control over what happens when
the value doesn't fit.