Allow functions to throw errors :^)
Functions can now have a trailing "throws" keyword before the return
type is declared. This is syntactic sugar for ErrorOr<T> where T is
the return type.

Errors are thrown with a "throw <expression>" statement.

Calls to functions that throw will now be automatically wrapped in
C++ TRY(). This causes errors to automatically bubble to the nearest
caller.

At present, there's no way to stop the bubbling, so all thrown errors
will bubble out to the runtime's main function and abort the process.