Fix infinite loop while parsing illegal function call arguments
When the parser tries to parse an expression from "}" (for the second
argument of the function), it errors out and doesn't increment the token
index. But because the inner loop that repeatedly parses function
parameter expressions never considered errors, it would get stuck in an
infinite loop failing to parse "}".

The solution is simply to check for errors and bail out of the loop if
there is one. A regression test has been added to the parser tests.