typechecker: Allow assignment of T to T? without Some on type checking pass
This feature was present in some tests and was broken by the previous
commit due to type unification happening on more branches of the
`typecheck_expression` function.

The typechecker now has a check for this specific case and returns no
error instead of type mismatch.

This commit also adds tests for this behavior and fixes old tests that
previously passed, but shouldn't have, as they returned an `i64` with an
`i32` type hint.