typechecker+tests: Implement return analysis
This commit adds return analysis to the type checker. After each
statement in a block is checked, we check whether the statement we just
analyzed definitely returns. If it does, then we also mark the current
checked block as definitely returning. This allows us to propagate the
"definitely returns" information up to the function checker, which can
then verify that the function returns if it's not of a void type.