Add basic integer overflow detection to integer addition
As per the README, we want to make integer overflow or underflow a
runtime error. This commit is the first step in this direction, as it
adds overflow detection and a hard error (VERIFY) to integer addition.
The implementation in the code generator is slightly hacked, as we can
only simplify the match statements once we have the checked arithmetic
functions for all operations. But nevertheless, it detects both integer
overflow and underflow for all integer types.

Note that due to typechecker bugs/incompleteness, we can generate code
that adds integers of different sizes. That's what would happen in the
new underflow test if it weren't for the literal cast. The usage of
__checked_add makes this a C++ compilation failure due to template
substitution failure. As it is much more involved to check overflow with
two different types being added, along with the implied language
semantics (which of the two is the resulting type?), I believe this is
fine, especially because it doesn't regress any tests and the
typechecker should make the literal have the correct type anyways.