runtime+codegen: Put most of lib.h into the JaktInternal namespace
Instead of inventing a number of creative naming styles to prevent name
clashes with C++ generated from user code, let's just put everything
into a unambiguous namespace and use regular naming within that.

Unfortunately C++ is very specific about what's counting as a reserved
name in the global scope, so we can't make the namespace nearly as
obnoxious as I'd like to.