main: Use pico-args for better CLI
pico-args is a zero-dependency minimal argument parsing library that
just handles the bare basics of argument parsing. We use it here to
provide a basic CLI that's a little more ergonomic than before, plus
some help output. This is mainly preparatory work for subcommands,
specifying output files, etc. which require more sophistication than
pico-args.

The main two reasons for choosing pico-args is (1) it's super tiny and
fast, keeping the development cycle snappy even with full builds, and
(2) most argparsing things are handled in our own code, making it easier
to port to jakt later on. We'd just have to replicate the low-level
parsing facilities of pico-args, which are not too complex in comparison
to things that structopt or clap would handle.