typechecker: Infer return type from first return in function body
Previously, if the return type was unknown, we would infer the return
type from the last return statement in the top level block of a
function body. However, this would not account for child blocks, so
this example would be inferred as `void`:

```jakt
function foo() {
    if (true) {
        return 1
    }
}
```

Now, if the return type is unknown, and the function contains a return
statement anywhere in the function body, we infer the return type from
the first return statement in the function body. For example:

```jakt
function foo() {
    if (false) {
        return 1
    } else {
        return true
    }
    return "foo"
}
```

We will infer the return type to be i64 from `return 1`, as it appears
before `return true` and `return "foo"`, even though `return 1` is
unreachable.

This is because we cannot know ahead of time which blocks may be
executed, for example, if the if statement was `if (rand() % 2 == 0)`
instead of `if (false)`.