parser: Parser infinite loop when setting enum value to unclosed bracket
The parser for `parse_array` didn't bail when parse_expression failed
and it could not recover from the error, resulting in an endless loop.

Likewise, `parse_enum` didn't contain a loop itself but the fact that
it didn't error out when encountering Eof when expecting a } makes
outer-level parsers not aware that Eof has been reached, which makes
them continue their loops.

I'll be checking more functions that parse enclosed lists to make sure
they bail as well when the last token they meet doesn't match with the
token that finishes their enclosed list.