Errors: Catch errors that occur at the end of a line
Previously the error printer missed errors occurring at the end of a
line due to using a < instead of a <= check. This commit updates the
check and deletes a block of code that prints out errors that occur
at the end of a source code file as these errors are caught with the
updated check.