documentation: Add coding-style document
This is largely based on the SerenityOS CodingStyle.md, with many
sections removed, with examples converted to Rust, and an additional
section on error messages. I barely understand Rust at all so there are
very likely to be some areas I have overlooked, or some issues with my
code examples, but this should be a good start. :^)