codegen: Don't rely on ScopeGuard for defer
Since the language is built from scratch, why depend on C++ for defer?
This is a first cut and there are probably still some kinks to iron out.

This patch removes the ScopeGuard dependency and instead makes a
built-in version of defer. Currently this works by creating a Vector of
Function at the start of each function. Whenever a defer block is
reached, the block is pushed onto the Vector. Whenever the function
would return it is instead directed to the end of the function, where
the defers are handled before finally returning.

The main benefit of this approach is we're able to call fallible
functions inside of defer blocks. Current semantics are that
deferred blocks run at the end of function scope, not local block
scope. First every defer block is run, and then if any block
failed, the error of the first one to fail is returned.

So for the code

```jakt
function main() {
    defer println("foo")
    return 0

    defer println("bar")
}
```

this would be the output:

```sh
foo
```

And for

```jakt
function fail() thows {
    println("fail")
    throw Error::from_string_literal("boom")
}

function no_fail() {
    println("no fail")
}

function main() {
    println("1")
    defer no_fail()

    println("2")
    defer fail()
}
```

this would be the output:

```sh
1
2
fail
no_fail
Runtime error: boom
```

And for

```jakt
function main() {
    for i in 0..3 {
        println("{}", i);
        defer println("deferred {}", i)
    }
}
```

this would be the output:

```sh
0
1
2
deferred 2
deferred 1
deferred 0
```

Not sure if the last one should be considered a bug or a feature.