runtime: Create extern declarations for as_truncated and as_saturated
Already replace some of the integer cast tests to check that this is
working. The function and the builtin syntax currently generate the
exact same code; though we have to aid the function in deducing its
argument types. This is not a problem as both conversion operators won't
be used in such contrived situations in practice.