runtime: Add Array::contains method
We already have this for Sets/Dictionaries, and it seems like a
nice idea to generally have the same APIs where applicable across
different containers.

For now, this only works with floats/integers/strings, since
Traits<T>::equals is not defined for the other complex builtin
types / other classes and structs, but those can conceivably be
generated in the future.