Parser: Fix infinite loop while parsing illegal function parameters
Previously, the compiler would get stuck in an infinite loop when
trying to compile a function parameter list and encountering
something that's not a parameter due to the parser being stuck in the
loop once the error was encountered. An example of this would be
trying to use default parameters, which are not a part of the language
yet. This commit fixes that by breaking out of the loop once the error
is encountered and introduces appropriate test cases.