runtime: Rewrite String class to propagate errors everywhere
This snowballed quite a bit, but let's just get the pain overwith.
Basically, this patch removes the assumption that String construction
always succeeds and makes functions that allocate new String objects
return ErrorOr<String> (corresponding to "throws -> String" in Jakt.

We're now very detached from the SerenityOS AK::String API, but this
is definitely the way forward as panicking on allocation failure is
no good. :^)