Parser: Allow function parameters to be mutable
The function parameter parser tries to keep track of whether the
keyword mutable has appeared before a variable declaration. However,
it previously only used this information when considering the this
parameter. When considering any other parameter, it would instead use
the mutability information of the variable declaration. However, since
the parameter parser had already parsed the mutable keyword, the
variable declaration parser did not see the mutable keyword and
reported the variable as immutable. This commit fixes that by always
using the mutability information the parameter parser has gathered
in addition to the information that the variable declaration parser
provides. It also adds an appropriate test case.