meta+ci: Lint sample filenames
This is a simple check to make sure all samples follow using '_'
rather than '-' in their names. It's possible to exclude files
or directories from this check by adding the globs to
`SAMPLE_IGNORE_LIST' in the script. I've added "samples/apps/*"
to this list to start this off, as that may contain command-line
utilities where '-' is sometimes preferred.