runtime+codegen+tests: Support `continue`/`break` inside matches
Refactored `ExplicitValueOrReturn` -> `ExplicitValueOrControlFlow`
with `LoopContinue` and `LoopBreak` as new variants, and now there
are two macros that exist to handle those values: one with just
`return` (for codegen places outside loop) and one with `return`,
`continue` and `break`.

Codegen has been updated to track whether `continue`/`break` is
allowed in the context of the code that's being generated and whether
it should use a `return LoopBreak` or `return LoopContinue` when
generating `Break` and `Continue` statements, since matches run through
lambdas and we have to return that control flow information to the
runtime.