editors/vscode: Cache line break positions for faster span conversion
This whole situation is a mess, since LSP requires that we provide
file offsets and column indices in UTF-16 "coordinates". This patch
doesn't improve the architecture, but instead makes a cache of line
break locations in the document text so we don't have to scan from the
start of the document every time we want to convert between systems.

This significantly reduces UI lag caused by the type inlay hints.

The *nice* solution to all this would be that vscode adds support for
UTF-8 offsets in the future. There's a new LSP extension for this,
but so far they only advertise UTF-16 support.