selfhost: Store visibility on fields and methods
This is a deviation from the bootstrap parser, by making ParsedField and
ParsedMethod separate structs, rather than adding a Visibility field to
the existing ParsedVarDecl and ParsedFunction structs, where it doesn't
really belong. It would be nicer if we could extend those structs but
for now, they are just included as fields.