selfhost: Prevent infinite loops due to unhandled tokens in blocks
Previously, when parsing a statement, if the parser encountered an
unknown token it would return garbage without advancing the index. This
lead to infinite loops namely when parsing blocks. This commit fixes
this by incrementing the index when generating a `Garbage` expression.

Before this commit, running `samples/basics/error.jakt` would result in
an infinite loop, now it produces the expected error output.