typechecker: Make child namespace contents available in the parent
This makes it possible to reference types from a child namespace in the
parent namespace.
To make this possible, we have to keep a placeholder structure/enum to
actually reserve the struct/enum ID, and fill it later when needed -
note that this entry's contents are not meaningful at this stage of the
typechecker algo, so we don't need to make it actually correct.