selfhost: Prevent infinite looping on error reporting
Increment the token index when finding an unexpected token on
operator, struct/class and function parsing.

To prevent spamming error messages, we now have a flag to know whether
an error has already been reported in the current state. This is a
simple hack to mimic the way the bootstrap compiler works and isn't
implemented everywhere it should.

This patch also changes some class/struct error messages to display
the correct declaration type.