editors/vscode: Collapse duplicate inlay hints
The compiler may emit duplicate type hints in some situations,
such as "foo | bar | baz" match cases.

This patch deduplicates the hints in the language server for now.
The nicer fix will be to not emit duplicate hints in the first place,
but let's leave that for the self-hosted compiler instead of making
big changes to the baseline compiler now.