selfhost/parser: Fix parsing of multiple postfix operators
The `parse_operand_postfix_operator` function would only attempt
to parse a single postfix operator, so if there were multiple e.g.
`index++ as? u32` or `Namespace::Variable++` only the first would be
consumed and the parser would then choke on further operators.

This was fixed by simply looping the process of trying to parse a
postfix operator and glomming it on to the current expression until
a token that is not a postfix operator is found and the expression is
returned.

The diff looks really nasty because git tries very hard to find lines
in common even when there are none because of indentation changes, but
essentially all that happened is that the code was indented twice (once
to turn it into a proper function, and twice to wrap it in a loop) and
a bit of code was added at the beginning and end.