selfhost: Add help to command line arguments
Passing `-h` will now print the possible flags and options similarly to
how the Rust implementation does.

The `usage()` and `help()` functions now return the usage and help
strings rather than calling `println()` and returning an integer. This
allows for the usage message to be printed to stdout when '-h' is passed
rather than to stderr. Ideally this would be done with global constant
strings as in the Rust compiler, however that is not currently
supported(?).