selfhost/parser: Avoid infinite loop in enum variants parser
The variants parser didn't advance the index when the variant that it
encountered was just a name, and it also couldn't differ between a
unclosed enum and getting EOF after closing it since it advanced the
index before breaking on `RCurly`. Now the `RCurly` check is done at
the end of the loop, and it is not skipped inside the loop.