typechecker+tests: Use the most specific scope id for unknown namespaces
`unwrap`ping isn't an option if one of the namespaces in the access
chain isn't found. Since `checked_namespaces` is emmited even when
the namespace wasn't found (so that typecheck/IDE can continue knowing
there was a namespace access), using the last scope id
from the last known namespace when the next one wasn't found.