tests/codegen: Fix invalid test
Test `codegen/control_flow_inside_match` was being skipped due to
using `out` instead of `output` inside the `Expect` specification.
Plus other issues like `&&`/`and` and control flow analysis not knowing
that `test_continue`'s loop is unreachable, so I added a `TODO` and an
explicit return with an invalid value so it fails if it get there due
to any other bug.