selfhost: Switch to using Parser::parse instead of `parse_namespace`
Now main does not need to explicitly instantiate a parser (and have
to construct it with the magic number zero of index) nor does it have
to explicitly make reference to the internal `parse_namespace`
function.

This change brings the handling of the parser within main in line
with how all the other pieces of the compiler are handled.