selfhost+main: Add cli options needed for godbolt integration
This adds the options that I need to make godbolt integration work.
These are mostly to specify various file paths for cxx-compiler,
clang-format, .clang-format, runtime and the output directory.

Specifically, it's useful to make the rs and selfhosted compiler
take the same arguments so you can view generated code side-by-side.

To make option a little less spammy, introduce a simple ArgsParser.
This isn't meant to be permanent, but it makes the main logic a little
simpler for now.