jakttest: Move selfhost & jakttest to custom build dirs
If any other file named `main.jakt` is compiled with no
extra arguments, its resulting binary will be `build/main`.
This will make `run-all.sh` think that selfhost is already
built.
By moving both selfhost and jakttest to a separate build
directory we reduce the chance of this situation happening.
I also made it a variable in case anyone wants a different
output directory, with the corresponding changes in Jakttest
to accept the custom executable path.