editors/vscode: Fix npm install
Change the script used in npm install for the vscode extension so as to
use && instead of ; as separators for the commands. Perhaps ; works on
linux (it doesn't on windows) but apparently semicolons as separators
means the commands that come after will be executed regardless of
whether the previous one failed or not, which doesn't seem correct
either.