selfhost/typechecker: Fix bogus check in NamespacedVar typecheck
To check whether a NamespacedVar was actually a simple enum constructor
invocation, we try to resolve it as a call. Of course, that resolution
can still fail, but we were checking if it failed by seeing if there
were *any* errors already logged by the compiler.

This meant that every simple enum constructor invocation would generate
an error once any other error had been seen.

Fixes a huge flood of errors when building the selfhost compiler with
itself.