jakttest: Make Jakttest not poll for processes
Instead of polling for every running test and spinning, we call
sigwait() when we don't have anything to run, so that Jakttest's process
is suspended until any child exits (SIGCHLD). When we wake up we do
generic polls with waitpid(-1, ...) to process all the tests that
finished executing and mark their directories as freed, launch more test
processes and go back to sleep.

This way we avoid spinning which gets in the way of the N jobs running
clang++ that are compiling the generated C++ from running the tests.