typechecker: Search for function names in unnamed children namespaces
This should be done also for structs/other namespaces, but searching for
a name should look inside the childen anonymous namespaces so they're
resolved correctly. This commit implements it for functions, using a
queue and a visited set. Can't use a hashset/array.contains() because
the concepts that the backing C++ structures require are not implemented
for our custom ID structures.