codegen: Prevent `typename typename Foo` from appearing in generated C++
RealBlueta came across this bug. Both `codegen_enum_match()` and
`codegen_type_possibly_as_namespace()` which it calls, add `typename `
before the type name itself, if the type was defined inside a
namespace, and it followed the `CheckedEnumVariant::StructLike` branch.
The other branches that output `typename ` are not affected, and
removing `typename ` from them causes tests to fail.

A proper solution would be to have one place for responsible for
outputting `typename ` but this works and doesn't break anything. :^)