parser+typechecker: Merge namespaces/imports when equivalent
If a namespace or import is added, it is checked against the
pre-existing namespaces/imports in the same (parsed) scope, and if there
is one that matches the same characteristics, it is merged with it. This
way we avoid creating two different scopes for the same namespace, which
results in false negatives, and we avoid creating redundant import
lists.