jakttest: Add Windows OSAL pieces for filesystem and process management
Add a CMake build rule for jakttest as well.

The file traversal and discovery code works, but the process spawning
code will remain "looks correct according to the docs" until there's
an alternative to run-one.sh. We can't run a bash script in powershell
with ./ the way Unix OS's do it.