selfhost/typechecker: Eval else condition on is op
This fixes a bug uncovered by AK on the following case:

```
enum E {
    A(x: i64)
    B
}

function main() {
    let e = E::B
    if e is A(x) and x == 5 {
        println("FAIL")
    } else {
        println("PASS")
    }
}
```

After the fix, the else statement gets evaluated when the
first condition is not true.

This works perfectly but it should be noted that it can
generate some redundant code during the code gen phase.

The redundant code probable gets optimized by llvm but
I will work on a more efficient solution in a later commit