codegen+runtime: Fix array slice offsets
`operator[]` and `at` were both applying the offset for a slice so the
values returned from the slice were incorrect for offsets other than 0

ArrayIterator `next` wasn't accounting for the offset when determining
if the end of the slice had been reached.

Added `slice_range` in `Array` to avoid subtraction underflow.