typechecker: Create typechecking step for namespace constructors
This allows a struct constructor to be created before the enum methods
are typechecked. This resolves #527 where it wasn't possible to call a
struct constructor in an enum method.

Also move `typecheck_function_predecl` to it's own step as function body
is typechecked in the generic case which needs to come after
enum/ struct constructor typecheck.