interpreter+tests: Add partial support for creating/returning optionals
This commit allows Some() to be used in comptime functions. Moreover,
they can be returned and used in regular functions as per usual. On the
other hand, None can be created, but returning it is problematic. Both
the typechecker and codegen work badly with None, as it does not store
any typing information.