typechecker: Fix method function_id with functions as parameters
Previously for structs a method was added to the module before any
functions as parameters but the function_id was calculated based on the
number of functions in the module after any functions as parameters were
added, resulting in a duplicate function_id added to the scope.

This would then result in the wrong `callee` being used when
typechecking a call to that method.