codegen+runtime: Fix pretty printing in exceptional paths
This commit replaces the previous pretty printing code with some
RAII based helpers for enabling/disabling pretty printing, along
with increasing/decreasing the indent leve.

The means that even if a TRY() fails in a function the pretty
printing will be returned to a valid state thanks to C++ destructors.