typechecker: Accept unknown type in various places
If the condition has an unknown type we'll accept it as well instead of
throwing an error, since we don't really know what the type of the
expression would be. Same goes with forced unwrap and for loop iterator,
where we don't know if what we're looking for actually would typecheck
or not.

This way we don't clutter the error output and let the user know that
we're stuck at the bad lookup.