typechecker: Add FunctionGenerics class to store specializations
`CheckedFunction` now uses `FunctionGenerics` to store generic params
and specializations. When a new `CheckedFunction` is created for a 
specialization it has a reference to the same `FunctionGenerics` 
instance as the base definition of the function.

This avoids infinite recursion when specializing a generic function
that calls itself as the function will only be specialized once
for each set of concrete types.