typechecker: Use function scope as parent for generic specialization
As specializing a generic function creates a new scope in the current
module, a function from another module is no longer able to make local
calls within that module.

This change instead uses the original function scope as the parent when
creating a new scope for specialization, allowing it to make calls local
to the module it is in.