jakttest: Better output for test failure cases
Previously we would only look at the output based on the expectation, so
if an error was expected and the test had no errors the output wouldn't 
be shown. 
Also the opposite where if output was expected and there was instead an 
error.

Also improves the output for `ErroredAtEarlierStage` and adds 
`ErroredAtLaterStage` for a case for example if you are expecting an 
error and instead get a cpp error.