runtime: Use correct size bound for array slice iterator
This also changes the behavior so that the bounds are determined only
when the slice is created. So that creating the slice with no range end
bound will create the slice to the end of the array but not also include
elements added to the array later in the slice.