selfhost: Set default runtime path based on location of jakt binary
Now we look in current_executable_directory/../include/runtime for the
runtime files instead of $PWD/runtime by default. This makes it so that
one doesn't have to pass -R in most cases. Windows is still a bit iffy
because the FilePath utility doesn't understand that Windows paths have
backslashes instead of forward slashes.