selfhost: Implement basic String Interpolation support
This commit adds basic String Interpolation functionality to the
self-hosted jakt compiler. The syntax used for interpolation is an
expression wrapped in `${}`. For example:
let foo = 123
println("foo is ${foo}")

There is full support for format specifiers supported by
StringFormatter:
let answer = 42
println("The answer in octal is ${answer:o}")

This makes string interpolation more distinguishable from the template
placeholders used in the string formatter (`{}`) this should avoid any
ambiguities before they even occur.

The `${expr}` syntax was chosen as it had the most upvotes in the
discord discussion about the topic.

In code generation, the formatting is implemented using
`String::formatted` to support all the known format specifiers.

Interpolation in the interpreter is implemented by calling the
`format()` prelude function to yield a JaktString.