codegen: Wrap enum variant call in enum type
Previously instantiating a Tuple with an enum variant as a variant would
give a c++ error for no viable constructor.

This is because for example:
`let tuple = (Foo::A, 1)`
would codegen to
`Tuple{ typename Foo::A(), static_cast<i64>(1LL)}`

Which would be a constructor of Tuple(Foo, long) with an argument of a
Tuple<Foo::A, long>
With the compiler not knowing how to convert a Tuple(A, long) into a
Tuple(Variant<A, B>, long)

It now will codegen to:
`Tuple{ Foo { typename Foo::A() } , static_cast<i64>(1LL)}`

Thanks to Andrew Kaster for explaining the c++ issue and working out
the fix.