all: Add `else` bindings on enum match
This allows us to avoid writing representative mapping for enum variants
by allowing the default mapping to be done in the `CatchAll` match 
pattern.

`CatchAll` with bindings is always typechecked last and will typecheck 
all missing variants with those bindings and codegen each one 
separately.

For example things like
```
match this {
    Boolean(span) => span
    NumericConstant(span) => span
    QuotedString(span) => span
    ByteConstant(span) => span
    CharacterConstant(span) => span
    ...
}
```
can become
```
match this {
    else(span) => span
}
```

This also allows for a mix of specific cases and a catch all such as
```
function type(this) -> TypeId => match this {
    Boolean => builtin(BuiltinType::Bool)
    NumericConstant(type_id) => type_id
    EnumVariantArg(arg) => arg.type_id
    IndexedExpression(type_id) => type_id
    IndexedDictionary(type_id) => type_id
}
```
becoming
```
function type(this) -> TypeId => match this {
    Boolean => builtin(BuiltinType::Bool)
    EnumVariantArg(arg) => arg.type_id
    else(type_id) => type_id
}
```