jakttest: Support cppincludes directive in Expect clauses
This will add each space separated directory as a -I directive to clang.

This is particularly useful if a submodule of the project has a native
header file living next to it in the source tree. In the future this
should be handled by the compiler invocations in the compiler itself,
but for now jakttest still calls out to a script for the cpp compile
step.

Co-Authored-By: Jesus Alvarez <jeezusjr@gmail.com>