codegen: Move match else case to the end for generic/value match
This allows an else case to be placed at any position in a value match
and only match if no other case matches. The other cases will still be
matched in order.

This makes the behavior closer to that of an enum match with the else
case only matching if no other case matches.