runtime: Add JaktValue builtin to represent different types of Jakt
This type can be used to let functions take arbitrary types as input
and match on them to decide what to do. This probably has many usecases
but the main one I have in mind is for string interpolation where
different types can result form arbitrary expressions and we want to
pass those to a formatting function.

The cpp implementation is generated from the enum in the prelude, but
since the prelude gets ignored in codegen we just provide the
implementation. In a perfect world the header file would be
automatically regenerated from the Jakt code but I have no idea how to
make this work using CMake :)