tests: Fix call to private constructors when testing private fields
These tests should ensure an error is thrown when accessing a private
field, however they both called private constructors in doing so. This
works for now since the error that is checkes is still thrown along
with a whole bunch of others, but this may change in the future.

The auto-private constructor for structs also was never explicitly
tested so this adds an explitcit test for that function