types: Add CheckedFunction.signature_matches
This lets you easily check if a typechecked function matches another in
signature. This means the function has the same name, number of generic
parameters and the same arguments where arguments that are not a
generic type need to match the other functions type exactly while
generic parameters always are the same between functions.