formatter: Extend conditional statements that use logical breakpoints
Extend from `if` to `match` `for` `while` `if` `guard`
Introduce `in_condition_expr_indented` so that there is no double indent
if the statement itself was already indented.

Additionally once using a logical breakpoint it will break on every
`and` and `or` which in general becomes a lot neater.

There are some cases to be handled later on mixing logical and non
logical breakpoints when it would be ideal to change this.

So for example:
```
guard long_condition and long_condition and short_condition {
```
would become
```
guard long_condition and
    long_condition and
    short_condition {
```
Even if the last 2 conditions could fit on the same line