types: Allow for overload sets in a scope
An overload set is a list of functions that all have the same name.
Previously a name could only represent a single function. This change
makes it possible to change this in the future although currently it is
asserted that only a single function can be added to an overload set