typechecker & codegen: Allow for function overloading
This adds the possibility to overload functions with the same name but
different signatures. For this, the namespace no longer contains simply
a mapping from names to function, but to a list of functions. This
list is iterated when typechecking a call to find the most relevant
implementation of the function based on the signature by using
the implementation with the highest total specificity of all parameters
that matches the call signature.