editors/vscode: Add durationLogWrapper
Before the `console.time` calls did not give expected results since they
were used in an async setting — they did not include the awaited time,
which tells how long the compiler uses on the different actions.

This patch also adds a `durationLogWrapper` which also takes all the
different return paths into account (would be nicer if TypeScript had
a `defer` keyword! :^) ).

Lastly, a `getClickableFilePosition` is added
which servers to purposes.

1. It make easy to see where the current action is called since you
can `CMD + click` on the file path (at least on MacOS).
2. Differentiate between multiple calls of the same action since the
`console.time` requires that the last label if finished before a new
one is called.