selfhost+parser: Add relative keyword for imports
If "relative" is used with import, as in,

import relative foo::bar

The compiler will search for the module in a sub-path relative to the
file.

If "parent::" is prepended to the module name, as in,

import relative parent::foo::baz

The compiler will search "up" a number of paths for the module.

The number of parent paths to search relative to the file can be
provided using syntactic sugar: "parent(n)::".

"parent()::" is the same as "parent::".

"parent(3)::" is the same as "parent::parent::parent::"