parser: Initial support for capturing `this`
This patch allows us to capture the `this` raw reference in a closure,
such as in:

```
struct Foo {
    message: String

    function bar(this) {
        let baz = function[this]() => println(this.message)

        baz()
    }
}

function main() {
    let foo = Foo(message: "WHF")

    foo.bar() // prints "WHF"
}
```

Previously we had pass in the enclosing type as a reference parameter to
achieve this.

For now, we do a by-reference (in C++ terms) capture of the current
object, but since we currently do not allow storing or returning
closures, this should be safe.