Everywhere: Add support for 'default' binding clauses in match patterns
Sometimes two patterns can share the match body...except one is lacking
a required binding. Enter default bindings, where you can specify the
missing ones and give them a value to be used in the match body:

    enum Foo {
        A(a: i32)
        B
    }
    ...
    match foo {
        A(a) | B default(a = 42) => a + 1
    }

This makes the following pattern unnecessary, which means we can also
add a diagnostic to ensure that all OR'd patterns bind the same set of
names.

    let a = 42
    match foo {
        A(a) | B => a + 1
    }