runtime: Return the correct types from Dictionary::get()/operator[]
A friendly class type cannot be used in dictionaries otherwise.

The issue stems from the fact that HashMap<T, U> hands U out as
`Traits<U>::PeekType` (and not U itself), so we can't go back to the
original U object from that if they're not the same type.

Note that the HashMap part of this *should* (and will) be backported to
Serenity, so further conflict with AK is not a big issue.